Fix #3264: Reset generatedSAMWrapperCount to 0 for nested classes.
When generating nested classes through `nestedGenerateClass`, the
scoped var `generatedSAMWrapperCount` was reset to `null`, whereas
it is set to `new VarBox(0)` for top-level classes. This could
cause NPEs in `synthesizeSAMWrapper` if an LMF-capable SAM function
was nested in the body of an anonymous JS class.

This commit fixes the issue by resetting `generatedSAMWrapperCount`
to `new VarBox(0)`.