Move things from `linker.*` to `linker.standard.*`.
We move `Linker`, `LinkedClass`, `LinkingUnit` and `Versioned` to
the `standard` subpackage of the linker, since they will not be
part of the super-stable API.

In addition, we rename `Linker` to `StandardLinkerImpl` and only
publicize its companion object. This will be important when we
rename `GenLinker` to `Linker`, otherwise the recommended imports
of `linker._` and `linker.standard._` would produce an ambiguous
import of `Linker`.