Fix #3267: Always use the enclosing class type for this local vars.
The fix for #3058 in 1518b704802ea95c3521831d91610905f022ca25 was
wrong. It made things worse for cases where the self type S of the
enclosing class C is not a subtype of C.

It turns out that, although scalac *declares* the `_$this` local
variable with the self type S, it is in fact *used* everywhere as
if had the type C. This is not very surprising, given that it
compiles it down to `this` in the bytecode (yes, it *reassigns*
`this` for tailrec methods).

In this commit, we therefore ignore the declared type of `_$this`
and always patch it up to C, the enclosing class type. This means
we need to adapt the right-hand-side of the initial declaration,
and of reassignments in calls to the tailrec label. Everywhere
else, a `This()` is used, which naturally translates to having the
type C.