Unseal `ModuleKind`.
We will want to add new module kinds in the future, so that class
cannot be sealed. We make its constructor private so that it still
cannot be extended from the outside.

This means that our own code matching on those will not receive
exhaustivity warnings, however, which is a pity, but seems
unavoidable.