Convert all the RuntimeLong tests to Long tests.
`RuntimeLong` is not really part of our public specification, so
it does not make much sense to directly test it. In the future, its
IR should even be a resource of the linker, in which case the tests
would not even compile because `RuntimeLong` would not accessible.
Moreover, testing `RuntimeLong` is not sufficient to test
alternative implementations of `Long`s, nor to test all the crazy
stuff that the optimizer does with `Long`s anyway.

In order to prevent the compiler and/or the optimizer to simply
constant-fold everything away, all the operations are hidden behind
`test` methods that selectively "show" and "hide" constants from
the optimizer.

Overall, the changes in this commit improve the coverage, since
they expose more conditions to the optimizer.