Alternative implementation of `Long`s on top of ES `BigInt`s.
This commit introduces a new way to emit `Long`s in the emitter,
using the upcoming ECMAScript `BigInt`s. Although they are
arbitrary precision integers, rather than 64-bits, they are
designed to allow implementations of 64-bit integers on top of
them.

The alternative implementation can be selected with the option
`allowBigIntsForLongs` of the `ESFeatures`.

As of this writing, only a very recent V8 can run the produced .js
files, and only with `--harmony-bigint`. Therefore, the present
commit does not any automated tests to the CI, which makes this
feature even more experimental. The implementation can be locally
tested with https://github.com/sjrd/scala-js-env-d8, using a
locally built d8 (the V8 debug shell).

The general encoding is straightforward: `Long`s values are encoded
as `bigint`s whose value fits in the signed 64-bit integer range.
The operations translate as follows:

* `a + b` -> `BigInt.asIntN(64, a + b)`
  (similar for most operations)
* `a << b` -> `BigInt.asIntN(64, a << BigInt(b & 63))`
  (similar for `>>`)
* `a >>> b` -> `BigInt.asIntN(64, BigInt.asUintN(64, a) >>
  BigInt(b & 63))` (because `BigInt` does not have logical shift
  right)