Migrate to the Pipeline feature of Jenkins 2.
We replace our makeshift `matrix.xml` + Groovy scripts inside the
Jenkins config by a single `Jenkinsfile` script. It defines both
the matrix, and how to use to create a pipeline of jobs.

This `Jenkinsfile` is intended to be used with a Multibranch
Pipeline (https://jenkins.io/doc/book/pipeline/multibranch/). It
contains the logic to run a subset of the configurations for PRs
(and push'es to branches), compared to the nightly builds.

Since it is unclear how to distinguish "nightly" from "weekly"
builds using this system, we take the opportunity to go back to
building both the former 'nightly' and 'weekly' matrices as one
'full' matrix. We also rename the 'pr' matrix into 'quick'. Lately,
weekly builds have been faster nightly builds anyway.

Using a groovy script instead of XML to define the matrix allows
to do it programmatically, avoiding some repetitions and improving
readability.