Fix #3295: Do not set __ScalaJSEnv.javaSystemProperties in the sbt plugin.
Instead, we set it up in our own build.

That feature was initially included in the sbt plugin because our
build needed it, but it has always been a bit hacky. In particular,
it requires the sbt plugin to know too much about how JS files are
executed, notably whether they are scripts or modules. Moreover,
that feature was the only reason left that the sbt plugin needed
to know about `__ScalaJSEnv` at all.