Blacklist pos/t6367.scala, because it spuriously fails too often.
This test stress-tests the type checker by design. On our CI, this
fails too often with weird errors (out of memory?). Moreover, it is
so bad that it seems to cause other tests running concurrently to
also mysteriously fail every once in a while.

Blacklisting that test should significantly reduce spurious
failures of partest runs.