Introduce the new API of ESFeatures instead of OutputMode.
This is done in a backward source and binary compatible way, so
`OutputMode` remains as the "source of truth" while `ESFeatures`
builds on top of it to provide the new API.

Whereas `OutputMode` was a sealed abstract class with case objects,
`ESFeatures` is an option bag of individual ECMAScript features.

Unlike the `OutputMode` API, it will be possible to evolve the new
API in backward compatible ways, which allows to lift it to the
stable API in `linker`, instead of `linker.standard`.