Blacklist LongTest.scala in FullOpt.
`LongTest.scala` uses optimizer-based generative programming to
generate a *lot* of tests. Worse, the way it optimizes `Long`s
turns that huge amount of test code into an *insane* amount of .js
code within methods. Closure then completely loses it with that
much code and all the intra-method optimizations it can do about
it.

Blacklisting this test when compiling for `FullOpt` dramatically
improves the time it takes to fully optimize the test suite
(between 4x and 5x on my machine).