Fix #3281: Backport the new `genConversion` from 1.x.
Starting with Scala 2.13.0-M3, if the typer needs to *adapt* an
`Int` to a `Long` inside a `match`, typically after overload
resolution has taken place, we end up with a tree of the form
`(x: Int).$asInstanceOf[Long]` in the back-end. This kind of node
is handled by the `genConversion` method, which was previously
buggy for the `Int`-to-`Long` conversion (among others).

This commit backports the implementation of `genConversion` from
master, which was rewritten in
9890ee769ee55adc0ef0514760d08e3fe3a3c2a0. We needed to adapt the
code to rely on `TypeKind`s instead of `jstpe.Type`s, because in
0.6.x the latter do not know the difference between the various
sub-`Int` types.