Get rid of the `Continue` IR node, as well as labels on loops.
Where we previously used `continue` statements in the IR, as in:

    lab: while (cond) {
      ...
      if (c)
        continue lab;
      ...
    }

We now use a labeled block inside the loop, as well as a return off
that labeled block, instead:

    while (cond) {
      lab[void]: {
        ...
        if (c)
          return(lab) undefined;
        ...
      }
    }

This allows two important simplifications in the IR:

* We do not have the `Continue` IR node anymore.
* `While` and `DoWhile` loops do not have labels anymore.

With those changes, labels are only used by `Labeled` blocks and
their corresponding `Return` statements.

To prevent some regressions in the quality of the generated .js
code, we recover JS `continue` statements out of the above
structure in the `FunctionEmitter`. We also received some
*improvements* for free, because the `tailPosLabels` can kick in
for those labels.