Make the label of `Return` nodes mandatory.
A `Return` with a `None` label implicitly meant returning from the
enclosing `Closure` or `MethodDef`. In the spirit of solving all
problems with `Labeled` blocks, we can get rid of this special-case
by introducing an explicit function-wide `Labeled` block when
necessary.

Instead of emitting:

    def foo__I(): int = {
      ...
      if (c)
        return 5
      ...
    }

we would now use

    def foo__I(): int = {
      _return[int]: {
        ...
        if (c)
          return@_return 5
        ...
      }
    }

which is equivalent.

This complicates somewhat the compiler back-end, since it now needs
to take care of adding that label if necessary (it could *always*
add one, but that would not be nice, and defeat most of the
purpose of this change), and it does not really simplify anything
later in the pipeline either.

Moreover, we observe a slight regression of quality of generated
.js code in some rare cases: when there is a `return` in a
`void`-returning method, and that method is not inlined, the
`FunctionEmitter` cannot replace the `break _return` by a `return`
in JS. This is due to a technicality in `FunctionEmitter`, and this
restriction could be lifted in the future.

The main benefits are more regular IR, as well as a potential
performance improvement in the optimizer. Indeed, previously, every
time we inlined a method, we needed to preemptively summon a lot of
complicated machinery to introduce a label in case the body
contains a `return`. In most cases, it doesn't, and we detect after
the fact that the label is useless and can be removed. Instead, we
can now straightforwardly process the body without any extra
action. In the rare cases where the body did contain a `return`,
the optimizer will encounter the `Labeled` block, and spawn the
machinery then. In all the other cases, the machinery is completely
by-passed.