Fix #2751: Fix the definition of `js.Date` for invalid dates.
For invalid dates, created from a string that cannot be parsed, the
methods of `js.Date` return `NaN`. Therefore, they must be typed as
`Double`s. By extension, this covers up for Firefox returning -0
for some valid date as well, instead of +0 like everyone else.

In the process, we fix the behavior of `j.u.Date.parse` for invalid
inputs.