Fix the way the bootstrap test prepares `testSuite/test:fastOptJS`.
Since b9ec4ee06a26b895f168c3ecdbae5b9724beb447, it is basically
invalid to call `testSuite/test:fastOptJS` if
`scalaJSStage in testSuite` is `FullOptStage`, and conversely. It
didn't cause any hard failure, but was ill-advised since the
blacklisting would not be applied appropriately.

The bootstrap test was erroneously doing just that. This commit
fixes that test, and adds a check in the test suite settings to
make sure no one tries to do that anymore.