Fix #2383: Rewire Throwable to extend JavaScript's Error.
In order for our Throwables to be truly recognized by JavaScript as
error classes, they must extend `Error`. With ECMAScript 2015
classes, this is possible (and easy), but with ECMAScript 5.1
functions and prototypes, we cannot *truly* make them errors,
because there is no way we can convince ECMAScript to give them an
`[[ErrorData]]` internal slot. That is because the `Error`
constructor will only do that with a non-undefined `new.target`.

To minimize the gap between ES 5.1 and 2015, we still rewire the
prototype of `Throwable` to that of `Error`.

In both cases, this rewiring means that `Throwable` does not
inherit the methods of `j.l.Object` anymore, so we need to
explicitly introduce bridges for those in `Throwable`.