Move the default value of some sbt settings to the Global scope.
We do this for `scalaJSLinkerConfig`, `jsEnv` and
`jsExecutionFiles`. It can be convenient to override those settings
at the global level in a build, for example to globally enable or
disable batch mode, or switch the JS environment.

The only Scala.js-specific settings left that we initialize in the
project scope are `scalaJSModuleInitializers` and
`scalaJSUseMainModuleInitialize`. It does not make sense to set
those at the global level, since they depend on the specific
codebase being linked.