Fix the reported JS names in conflicting overloads messages for good.
We were previously guessing the JS name to display based on the
symbol's name. We tried fixing it in
59eb62cf85701385a1a73af7a5ee198091a13703, but that was not good
enough. In general, for exports, there is no correct answer to be
found. This is pretty obvious once we take into account the fact
that a method can be exported several times under different names.

Therefore, in this commit, we stop trying to guess the name, and
instead pass it down from the point where we actually know the JS
name for which overload resolution could not be solved.