Support Scala 2.13.0-M4 in all the non-tools artifacts.
This commit puts the codebase in a state where we can build the
compiler plugins, libraries and test suite with Scala 2.13.0-M4
and its new collections. The tools API and related artifacts are
not yet updated.

This is the minimal set of changes necessary to be able to build
and back-publish for 2.13.0-M4 when it comes out. We will not be
able to publish tools artifact for that version, but that is not
critical.