Fix #2050: More accurately test flushing of `OutputStreamWriter`.
There is nothing that specifies that `flush()` itself must be
called to flush the underlying `OutputStream`. `OutputStreamWriter`
can instead rely on the fact that calling `close()` is supposed to
effectively flush the underlying stream as well.

Instead of testing the `flushed` flag of our
`MockByteArrayOutputStream`, which is designed to test whether
`flush()` itself has been called, we verify that the bytes have
actually made their way to the mock output stream.