Impose the total ordering for Floats and Doubles in j.u.Arrays.
Starting from Scala 2.13.0-M5, Floats and Doubles do not have a
unique `Ordering` anymore. They have a `TotalOrdering` consistent
with `java.lang.{Float,Double}.compare`, and an `IeeeOrdering`
consistent with the comparison operators. The one we want in
`java.util.Arrays` is the total ordering, according to the Javadoc.

Since there is no way to easily get the total ordering across
versions of Scala without warning, we simply redefine them as
private implicit objects in `Arrays.scala`. The implementation is
short and trivial, so that is easy to do.