Optimize switch-matches that have 0 or 1 case.
They come up in reasonable source code that use `match`es as a
convenience for a sequence of `if`s, or with patterns such as

    someComputation() match {
      case -1 => doOneThing()
      case n  => doAnotherThing(n)
    }

The simplification is done in the compiler rather than the
optimizer, because it helps the optimizer to directly receive the
`If`s or `Block`s. Moreover, there is no optimization that the
optimizer could do that would turn a non-optimizable match into an
optimizable one, so we're not losing any optimization opportunity
by doing it only once in the compiler.