Complete rewrite of java.util.Formatter.
This class was written a very long time ago, when quality standards
were not as high as they are now. In this commit, we entirely
rewrite it while aiming for:

* Readability
* Performance
* Fidelity in a lot of corner cases wrt. the specification on the
  JVM, in particular in error cases

The corresponding test suite is significantly enhanced.

In the process, this commit fixes #1619 (thousands grouping), as
well as many other issues, among which:

* Formatting of astral characters with `%c`.
* Formatting of `null` with all the format specifiers.
* Handling of precision in non-numeric strings (basically
  `substring`).
* Capturing `IOException`s thrown by the target `Appendable`.

Despite the large number of fixes, the following items are still
not addressed:

* #3202 and, more generally, various discrepancies between the
  formatting of floating point numbers wrt. the specification,
  with `%f`, `%e` and `%g`.
* Hexadecimal format for floating point numbers with `%a`.
* Formatting of date/times with the `%t.` family.
* Formatting of `BigInteger`s and `BigDecimal`s.