Fix malformed input length for lone high surrogate in UTF-16.
If a high surrogate is not followed by a low surrogate, our
implementation reported a malformed input of length 2 (counted in
bytes). On the JVM, this reports a malformed input of length 4,
hence also swallowing the next 2 bytes, even if they could be
interpreted as the beginning of a valid UTF-16 sequence.

This commit is a step towards fixing #2041.