Fix malformed input length for too-long sequences in UTF-8.
If a multi-byte UTF-8 sequence encodes a code point that could have
been encoded using a smaller sequence, it is an error. Our
implementation reported a malformed input of length N in that case,
where N is the length of the sequence. However, the JVM
implementation always reports a malformed input of length 1 for
this error. This commit fixes the discrepancy.

This fix is the last step to fixing #2041.