Fix #3426: Fully support lazy vals in non-native JS classes.
This turned out to be relatively simple. There is only a little
dance to correctly consider lazy *fields* as non-exposed, but the
lazy *def* that accesses it as exposed, both in 2.11- and 2.12+.

The commit also includes a better error message when trying to
include a `lazy val` in a JS trait. The code previously failed to
compile as well but with an obscure error message.