Fix #3406: Use the JS pattern and flags for group detection.
`Pattern` preprocesses the `pattern` string to extract some
Java-specific features, such as in-line flags, before giving the
patched patterns and flags to a `js.RegExp`.

Before this commit, the logic for group recovery in `GroupStartMap`
used the original pattern and flags. In this commit, we use the
already processed pattern and flags instead, so that this
preprocessing carries over to the regexes used for group recovery.