Support Scala 2.13.0-M5.
There is one non-trivial change: `js.ArrayOps` had to be completely
reimplemented for 2.13.0-M5, because the collection framework does
not accept collections classes whose `repr` is not themselves
anymore. This means that `js.ArrayOps` must be separated from the
collection *hierarchy* per se, and instead be an independent
`AnyVal` that happens to implement the entire same API. This is a
lot of code, but is similar to what `scala.collection.ArrayOps`
does for `scala.Array` (only that we have the API of `Buffer` to
support, in addition of just `Seq`).

These methods receive corresponding tests, since they are not
inherited from the upstream implementation anymore.