[no-master] Fix #3432: Use the new API of jsdom v10 in JSDOMNodeJSEnv.
jsdom v12.0.0 dropped support for the "old API" that was legacy
from v9. We were still using that old API so that we could support
jsdom v9 and v10 (and v11), but this is no longer an option for
v12.

This commit migrates to using the "new API" introduced in jsdom
v10.0.0, which brings support for v12.x.

In order to keep supporting v9.x, we use a dynamic fallback to the
previous implementation.