Handle cancellation correctly in 1.x's `ComRun`
In 0.6.x, calling `close` on the test adapter always kills the
underlying processes that are instantiated by the js environments.

Unlike 0.6.x, the latest 1.x test module doesn't kill the underlying
processes. This causes a resource leak that is fatal in Windows
environments (where the underlying process keeps a reference to the
output JS file indefinitely and prevents other tasks in the build tool
from linking again).

This pull request does the following:

1. Invokes `run.close()` in `ComRun`'s' close method.
2. Only runs `close` once, as `ComRun.close` can be called several times.

We're testing cancellation in Bloop for both Scala.js 0.6.x and 1.x and,
after these changes, all tests pass.