In `NoModule`, use true top-level `var`s for top-level exports.
Before, we needed to detect the global object at run-time (which
is not fool-proof) to add top-level exports as properties of the
global object. Now, top-level exports are truly defined as global
`var`s.

When using ECMAScript 2015 features, in theory we should use
global `let`s instead. However, that means that the variables
would not be available in the global object, and that poses a
problem for our test suite. Our test suite relies on the ability
to get an "exports namespace", which in that case would not be
possible. We *could* update our test suite with zillions of

    if (noModule) js.Dynamic.global.exportToTest
    else exportsNamespace.exportToTest

but that would be really annoying. Doing so is left for a future
commit.