Emit `let`s for top-level exports in ES 2015 `NoModule`.
This is trivial to implement, but annoying to test. `ExportsTest`
now duplicates every interaction with the "exports namespace",
since in ES 2015 `NoModule` the only way to access the exports is
as `global.TheExport`, which cannot be abstracted away. This
results in quite a bit of code duplication in the tests, but it
can't be helped.