Upgrade to Google Closure Compiler v20181210.
Besides benign API changes, the biggest change is that this new
version of GCC will not infer global variables that are accessed,
and instead raise errors if they are not declared. This means that
we need to track the set of *all* referenced global variables when
targeting GCC, rather than just the set of dangerous ones. We do
that with a new option in `InternalOptions`.

Because of too much lockdown in the ErrorManager API of Closure,
we cannot send the errors and warnings to our `logger` anymore.
Instead, they will be sent to the standard error, if any. Note that
in theory, this never happens.

This commit fixes #2815 as a side effect: the new version of GCC
does not exhibit the bug that caused that issue.