Fix #3181: Enable GCC when using ECMAScript 2015 features.
We update `ClosureLinkerBackend` and in particular
`ClosureAstTransformer` so that they understand and can translate
ECMAScript 2015 features.

Combined with the version of GCC, which understands ES 2015, this
allows us to enable GCC when using ES 2015 features.

The Jenkinsfile is updated to test more configurations, which
become meaningful due to this change.