Communicate top-level var decls as a list of var names in the emitter.
Before, it was an `Option[String]` where the string was JavaScript
code. This was not very friendly for a semantic treatment of those
declarations, notably in the Closure backend. Now we communicate it
as a list of the variable names, and let the top-level emitter take
care of turning that into appropriate JavaScript code.