Directly build the core JS lib as a `js.Tree`.
The core JS lib used to be written in a .js file with a custom
preprocessor to deal with variants in semantics, ES features and
module kind. In this commit, we directly generate it as a JS tree
instead.

Despite the relative loss of readability, this has a few perks:

* We can code-generate a few things that required a lot of
  duplication before.
* We can leverage a number of helpers in `JSGen`, which centralizes
  a bit more some of the encodings that we use.
* There is less duplication between similar code paths such as
  using ES 2015 or not.
* We can convert the AST into GCC's AST, eliminating the need for
  the awkward combination of text source code and ASTs.
* We can track the global refs used by the core JS lib itself in a
  more principled way.