Fix the name of jl_CloneNotSupportedException in inlineable init's blacklist.
The test accompanying this change does not actually fails before
the fix, but it does fail in a hello world. The bug cannot be
reproduced in our test suite: the existence of JUnit indirectly
refers to `java.lang.Enum`, which uses the secondary constructor
of `CloneNotSupportedException`, which supersedes the blacklist.