Fix #2108: Defer identifier name mangling to the emitter.
Previously, we banished JavaScript keywords in the IR, in several
places:

* The name of local variables
* The unprefixed name of fields
* The name of methods (not including their encoded signature)

This commit allows all JavaScript keywords (and special reserved
or meaningful identifiers such as `arguments`) in all these places.
For fields and methods, it was irrelevant anyway, since they are
always suffixed with something, and therefore the keyword is never
actually a keyword.

For local variables, we defer their mangling until the emitter, in
the pessimistic naming run (so we assume in the first run that no
local variable uses a reserved word). The optimizer helps by
eagerly renaming such variables on its own, to avoid the
pessimistic run. It has to do renaming for inlined stuff anyway, so
this is not more job for it. The pessimistic run is still necessary
in exported methods (which are not optimized) and if the optimizer
is disabled.

There is one place where the set of strict keywords must be avoided
in the IR, per semantics: in `JSGlobalRef`. Therefore,
`JSGlobalRef` does maintain a fixed set of illegal global variable
names, which cannot be accessed. They trigger compile errors.

There is still one thing which must be eagerly mangled: identifiers
starting with a digit (for fields, methods, and local variables).
That's because scalac thinks they are valid, but they cannot be
used at the start of an `IdentifierName` in JavaScript. We still do
not allow them in the IR, because the IR can reasonably follow one
of the two following rules:

* Either it follows JavaScript's rules, hence `IdentifierName`
* Or it follows the JVM rules, which allow virtually all Unicode
  characters (with a few exceptions such as . / < >).

Supporting the latter in the IR would mean much more preemptive
mangling in the emitter, which would slow it down for almost no
reason, since scalac uses none of those weird characters.

For that reason, the IR sticks to JavaScript `IdentifierName`s,
and it is the scalac back-end's responsibility to encode leading
digits away. We choose to prefix a leading digit with the character
`〇` (Ideographic Number Zero, U+3007, category `LETTER_NUMBER`),
which is a valid first character for an `IdentifierName`. This
avoids cascading manglings. At the end of the day, that character
will be Unicode-escaped as \u3007 by the pretty-printer to keep
the character set of the emitted .js as ASCII (this was already
implemented). If we had cascading manglings, it would be awkward to
specify interoperability between different languages compiling to
the IR.