Use ClassRef instead of ClassType in a bunch of places in IR trees.
This ensures that `ClassType` is only used when the referenced
class is a Scala class, i.e., it actually defines a *type*. For JS
classes, their type is `any` so it does not make sense to use
`ClassType`.

To be more consistent, we also use `ArrayTypeRef` in a few places
instead of `ArrayType`.