Erase `scala.runtime.BoxedUnit` to `java.lang.Void`.
This removes one more dependency of the IR specification on the
Scala namespace, in the handling of primitives, which is pretty
core.

`java.lang.Void` makes sense as the boxed class of `Unit` because

* `scala.Unit` is to `void` as `scala.Int` is to `int`
* Other boxed classes have a `TYPE` static field that represent the
  `classOf` of their corresponding primitive, and `java.lang.Void`
  precisely has a `TYPE` field that represents `classOf[Unit]`.

`java.lang.Void` is safe enough to reuse as it cannot have any
instance in the JVM spec (barring evil uses of reflection). Now
we're giving it exactly one instance: `undefined`.

Since `java.lang.Void` does not extend `java.io.Serializable`, this
introduces a weird discrepancy wrt. Scala/JVM: a boxed `()` will
not answer true to `.isInstanceOf[Serializable]`. However, it does
not seem like something would ever do. An alternative would be to
alter `java.lang.Void` and make it extend `java.io.Serializable`
anyway. This is a binary compatible extension so we could do it
later if necessary.

Another semantic difference introduced by this change is that
`((): Any).getClass().getName()` will now return `java.lang.Void`
instead of `scala.runtime.BoxedUnit`.