Include `RuntimeLong.sjsir` as a resource of the linker.
`RuntimeLong` is basically an implementation detail of the linker.
It shouldn't be part of the standard library. In the past, this has
caused troubles when we wanted to improve the encoding of longs,
as we had to keep backward and forward binary compatibility between
versions of `RuntimeLong`.

In this commit, we move `RuntimeLong.scala` in a separate project,
which is not published. Instead, its compiled products are stored
as resources of the linker. The fact that we have 3 variants of
the linker complicates things a bit:

* For the published JVM artifact, we use genuine resources.
* For the published JS artifact, we use source generation to embed
  the resources as constants.
* For the build itself, we cannot do either of those things, so
  instead we manually inject the products of `RuntimeLong.scala`
  on the `fullClasspath` of our projects, to enable linking.

The last bullet is concerning wrt. testing, as most of our tests
happen in the build itself, and we want to minimize environmental
differences between the build and the published linker. This issue
is somewhat addressed by the unit tests of the linker and the
smoke tests in sbt-plugin-test.