Include `RuntimeLong.sjsir` as a resource of the linker.
`RuntimeLong` is basically an implementation detail of the linker.
It shouldn't be part of the standard library. In the past, this has
caused troubles when we wanted to improve the encoding of longs,
as we had to keep backward and forward binary compatibility between
versions of `RuntimeLong`.

In this commit, we move `RuntimeLong.scala` in a separate project,
which is not published. Instead, its compiled products are stored
as resources of the linker.

* For the published JVM artifact, we use genuine resources.
* For the published JS artifact, we use source generation to embed
  the resources as constants.