Make division by 0 compliant (always).
So far, division by 0 has been an unchecked undefined behavior.
Contrary to the other undefined behaviors we have, this one has
always been problematic, because Scala as a language and a type
system do nothing to guard against it.

This commit makes division by 0 compliant, i.e., throwing an
`ArithmeticException`, in all cases. Some optimizations are thrown
in to avoid the additional checks when the rhs is a non-zero
constant.