Do not provide linked ClassInfos if there is a cycle
The currently provided ClassInfos are broken: The superClass /
interfaces / ancestors accessors may return null.

We could imagine "fixing" this by actually returning a cyclic object
graph. However, we'd have to be very careful to not continue linking
"too far" in this case: As soon as we try a member lookup we risk an
infinite loop.

Therefore, we simply provide the bare class names and fail fast. This
seems to be acceptable for such a bad error condition.