[no-master] Confine warnings in the compiler to Compat210Component.
As reported by Scala 2.11 and 2.12. There are some new warnings
emitted by 2.13 which we do not get rid of yet, notably the usage
of early initializers (which are required to deal with the scalac
API, AFAICT).

This commit is basically a backport of
895b833506a750ab34a3dd79a8b61099473a1247 from the master branch,
except that some warnings survive inside `Compat210Component`.